+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# rlang 0.3.0.9000
+# rlang 0.3.1
 
 This patch release polishes the new backtrace feature introduced in
 rlang 0.3.0 and solves bugs for the upcoming release of purrr
@@ -146,7 +146,8 @@ have plenty of time to change your code.
 * `as_closure()` wrappers now call primitives with positional
   arguments to avoid edge case issues of argument matching.
 
-* `as_closure()` wrappers now dispatch properly on methods defined in the global environment (tidyverse/purrr#459).
+* `as_closure()` wrappers now dispatch properly on methods defined in
+  the global environment (tidyverse/purrr#459).
 
 * `as_closure()` now supports both base-style (`e1` and `e2`) and
   purrr-style (`.x` and `.y`) arguments with binary primitives.